[PATCH] DVICO Fusion DVB-T1 Tuner (LG-Z201) fix
It is a small modification to the table that defines the way that the
LG-Z201 tuner is controlled for the DVICO Fusion DVB-T1 tuner card.

I believe that a mistake was made when the dvb tuner code was reorganised
(to use a generic table for the tuner information instead of inline code)
and as a result, the DVICO card doesn't tune properly.

The modification I have made to the table makes it behave like it did with
the old inline tuner code that worked.  The patch is on top of the 2.6.12
kernel.

Signed-off-by: Gregory B Frost <frosts1@hotkey.net.au>
Signed-off-by: Johannes Stezenbach <js@linuxtv.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>