[PATCH] ipmi: use CONFIG_DMI instead of CONFIG_X86
With Andi Kleen's x86_64 patch to use DMI, and my ia64 to use DMI, there is
now a new CONFIG_DMI option which takes the place of CONFIG_X86 to denote
the availability of the DMI functions.  Make the IPMI driver use CONFIG_DMI
instead.

Tested on ia64 2.6.15 kernel plus the previous patch, on a Dell PowerEdge
7250 Itanium2 server, and it now autodetects the IPMI KCS driver as
expected.

Signed-off-by: Matt Domsch <Matt_Domsch@dell.com>
Acked-by: Corey Minyard <minyard@acm.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>