V4L/DVB (3573): Cxusb: remove FIXME: comment in bluebird_patch_dvico_firmware_download
Removed the FIXME comment from bluebird_patch_dvico_firmware_download:
 /* FIXME: are we allowed to change the fw-data ? */
Yes, we are allowed. DViCO's Windows driver also does the same thing.
A single firmware image is used to support all of the bluebird boxes.
The firmware sets all devices to PID: d700. Instead of using that, the
driver replaces the d700 with the cold PID+1 before the download.

Signed-off-by: Michael Krufky <mkrufky@linuxtv.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab@infradead.org>