powerpc/mpc5200: fix build warnings on mpc52xx_psc_spi driver
The register definitions have been changed for the mpc5200 PSC ports
to cover some of the changes in the mpc5200b.  One change is that the
ccr register is now a u32 instead of a u16.  However, for the purposes
of this driver we want to continue to use 16 bit access to avoid
changing the existing (working) behaviour.

This patch allows the driver to continue to do 16 bit accesses without
the compiler complaining about it.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>