mtd: flash mapping support for Dreamcast VMU.
This patch adds support for the Sega Dreamcast visual memory unit as a
flash mapping. It requires changes in the maple bus driver (posted
separately) to support block reads and writes.

The VMU is a 'smart' flash device, with a built-in 8-bit controller -
for instance there is an erase before a write but it is hidden from the
user. But the device's overall behaviour means it works well with the
mtd layer and it is appropriate to add it as an mtd mapping.

Signed-off-by: Adrian McMenamin <adrian@mcmen.demon.co.uk>
Acked-By: David Woodhouse <David.Woodhouse@intel.com>
Signed-off-by: Paul Mundt <lethal@linux-sh.org>