s390/qdio: rename the misleading PCI flag of qdio devices
For each detected qdio device a line like to following is printed:
qdio: 0.0.4102 OSA on SC 1045 using AI:1 QEBSM:0 PCI:1 TDD:1 SIGA: W AP

The PCI flag is misleading as this stands for "program controlled interrupt".
Rename it to PRI "program requested interrupt" which is more accurate and
does not interfere with another popular piece of technology.

Leave the pci string in the code since changing that would result in a huge
patch.

Signed-off-by: Jan Glauber <jang@linux.vnet.ibm.com>
Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>